feta: updated prisma client to the latest version
paused for a while trying to figure out the problem causing failure of creation of an event!
