+++ package.json
@@ -14,7 +14,7 @@
     "@auth/prisma-adapter": "^2.4.1",
     "@googlemaps/js-api-loader": "^1.16.6",
     "@hookform/resolvers": "^3.6.0",
-    "@prisma/client": "^5.15.1",
+    "@prisma/client": "^5.17.0",
     "@radix-ui/react-alert-dialog": "^1.0.5",
     "@radix-ui/react-avatar": "^1.1.0",
     "@radix-ui/react-checkbox": "^1.1.0",
@@ -75,7 +75,7 @@
     "@types/react-redux": "^7.1.33",
     "autoprefixer": "^10.0.1",
     "postcss": "^8",
-    "prisma": "^5.15.1",
+    "prisma": "^5.17.0",
     "tailwindcss": "^3.3.0",
     "typescript": "^5"
   }
